Move DSE exceptions into their own package
