+++ SpiNNaker-data-specification/src/main/java/uk/ac/manchester/spinnaker/data_spec/Callable.java
@@ -16,6 +16,8 @@
  */
 package uk.ac.manchester.spinnaker.data_spec;
 
+import uk.ac.manchester.spinnaker.data_spec.exceptions.DataSpecificationException;
+
 /**
  * How to actually call a DSE operation.
  *
