+++ SpiNNaker-data-specification/src/test/java/uk/ac/manchester/spinnaker/data_spec/TestMemoryRegionCollection.java
@@ -23,6 +23,8 @@ import java.util.Set;
 
 import org.junit.jupiter.api.Test;
 
+import uk.ac.manchester.spinnaker.data_spec.exceptions.RegionInUseException;
+
 class TestMemoryRegionCollection {
 
 	@Test
